移動 / ガーデン監視機能を Farmer クラスへ移動
ガーデン関連の機能であるため移動
